requirements.txt
uvicorn==0.29.0
pydantic~=2.6.4
fastapi~=0.110.1
requests~=2.31.0
python-dotenv~=1.0.1
ollama~=0.1.8
sqlmodel~=0.0.16